get rid of containedLevelOps
